chore: resolve backport conflict in boot.py
