+++ frappe/boot.py
@@ -219,8 +219,11 @@ def get_user_pages_or_reports(parent, cache=False):
 	hasRole = DocType("Has Role")
 	parentTable = DocType(parent)
 
+	def exclude_disabled_reports(query):
+		return query.where(report.disabled == 0) if is_report else query
+
 	# get pages or reports set on custom role
-	pages_with_custom_roles = (
+	pages_with_custom_roles = exclude_disabled_reports(
 		frappe.qb.from_(customRole)
 		.from_(hasRole)
 		.from_(parentTable)
@@ -247,7 +250,7 @@ def get_user_pages_or_reports(parent, cache=False):
 		.where(customRole[parent.lower()].isnotnull())
 	)
 
-	pages_with_standard_roles = (
+	pages_with_standard_roles = exclude_disabled_reports(
 		frappe.qb.from_(hasRole)
 		.from_(parentTable)
 		.select(parentTable.name.as_("name"), parentTable.modified, *columns)
@@ -255,12 +258,7 @@ def get_user_pages_or_reports(parent, cache=False):
 			(hasRole.role.isin(roles)) & (hasRole.parent == parentTable.name) & (parentTable.name.notin(subq))
 		)
 		.distinct()
-	)
-
-	if is_report:
-		pages_with_standard_roles = pages_with_standard_roles.where(report.disabled == 0)
-
-	pages_with_standard_roles = pages_with_standard_roles.run(as_dict=True)
+	).run(as_dict=True)
 
 	for p in pages_with_standard_roles:
 		if p.name not in has_role:
@@ -273,7 +271,7 @@ def get_user_pages_or_reports(parent, cache=False):
 	)
 
 	# pages and reports with no role are allowed
-	rows_with_no_roles = (
+	rows_with_no_roles = exclude_disabled_reports(
 		frappe.qb.from_(parentTable)
 		.select(parentTable.name, parentTable.modified, *columns)
 		.where(no_of_roles == 0)
